Simplifying the algorithm to better react when opening Chrome DevTools.
ChromeDriver is implemented as a debugger for Chrome, and Chrome
currently only supports one debugger at a time (either DevTools or
ChromeDriver):
https://sites.google.com/a/chromium.org/chromedriver/help/devtools-window-keeps-closing

In practice, opening DevTools (when it is docked to the debugged window)
raises exceptions on the Java side.
This commit makes sure that those exceptions are ignored, so that,
instead of closing Chrome, the Selenium Java Robot is now paused until
DevTools are closed.

Note that if DevTools are opened in a separate window (instead of being
docked to the debugged window), the behavior is different: DevTools are
immediately closed because ChromeDriver reconnects and kicks out the
DevTools (there is apparently no way to detect and prevent this from
the Java side).