Adding a Chrome extension to make it possible to use DevTools.
This commit adds a Chrome extension which allows to plug the ChromeDriver
to a different tab than the one in which the API of the Selenium Java Robot
is injected. The extension makes the two tab communicate with one another.
The benefit is the ability to use DevTools in the tab where the Selenium
Java Robot API is injected for easier debugging.
As a result, calls to the Selenium Java Robot are much slower, so this mode
is only enabled when using the "chrome-debug" browser, not with the
"chrome" browser.